TDE Script
This stored procedure will create a master key if it doesn't exist.
An encryption certificate will be created with the password that is supplied.
The encryption certificate and private key is backed up to the location supplied (C:\ is an example).  Retain this for decryption purposes.
All of the databases are encrypted using the certificate and private key.